+++ changelog.md
@@ -3,8 +3,9 @@
 ## v2.8.6
 
 1. Added global MFA requirement to **POST** `/user/mfa/disable`
-2. Improved string table
-3. Improved config format
+2. Added member role sync-to-discord feature
+3. Improved string table
+4. Improved config format
 
 ## v2.8.5